rename: hospital/common 폴더 구조화
