+++ components/hospital/admin/icu-settings/order-color-settings/table-display.tsx
@@ -1,4 +1,4 @@
-import OrderTypeColorDot from '@/components/hospital/common/order-type-color-dot'
+import OrderTypeColorDot from '@/components/hospital/common/order/order-type-color-dot'
 import {
   Table,
   TableBody,
