+++ components/hospital/common/default-template-order/dt-order-creator.tsx
@@ -1,4 +1,4 @@
-import OrderTypeColorDot from '@/components/hospital/common/order-type-color-dot'
+import OrderTypeColorDot from '@/components/hospital/common/order/order-type-color-dot'
 import { Input } from '@/components/ui/input'
 import {
   Select,
