+++ components/hospital/common/default-template-order/dt-order-dialog.tsx
@@ -1,4 +1,4 @@
-import OrderTypeColorDot from '@/components/hospital/common/order-type-color-dot'
+import OrderTypeColorDot from '@/components/hospital/common/order/order-type-color-dot'
 import {
   Dialog,
   DialogContent,
