+++ components/hospital/common/order/order-title-content.tsx
@@ -1,4 +1,4 @@
-import OrderTypeColorDot from '@/components/hospital/common/order-type-color-dot'
+import OrderTypeColorDot from '@/components/hospital/common/order/order-type-color-dot'
 import { type IcuOrderColors } from '@/types/adimin'
 
 type OrderTitleContentProps = {
