+++ components/hospital/common/order/order-type-color-dot.tsx
