+++ components/hospital/common/patient/patient-overview.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function PatientOverView() {
+  return <div>PatientOverView</div>
+}
