+++ components/hospital/icu/main/chart/selected-chart/chart-body/table/chart-table-body/order-row-title.tsx
@@ -1,4 +1,4 @@
-import OrderTitleContent from '@/components/hospital/common/order-title-content'
+import OrderTitleContent from '@/components/hospital/common/order/order-title-content'
 import { Button } from '@/components/ui/button'
 import { TableCell } from '@/components/ui/table'
 import { OrderStep } from '@/lib/store/icu/icu-order'
